Added Mordimer init config

+++ src/Dto/RaceConfigDto.php
@@ -9,7 +9,7 @@ class RaceConfigDto
     public function __construct(
         public readonly array $asi,
         public readonly array $languages,
-        public readonly int $speed,
+        public readonly float $speed,
         public readonly int $darkvision,
     ) {
     }
